+++ CHANGELOG.md
@@ -13,7 +13,9 @@ title: Changelog
 * `<cc-map-marker-*>`: fix the way we document readonly properties (anchor, size, tooltip)
 * `<cc-zone-input>`: fix type for `_hovered` (string|null)
 * Storybook: fix logo URL
-* `<cc-env-var-form>`: add JSON mode
+* `<cc-env-var-form>`:
+  * add JSON mode
+  * add context for add-on configuration provider
 * `<cc-env-var-editor-expert>`: keeping only valid values when leaving expert mode
 * `<cc-env-var-editor-simple>`: add a strict mode
 * `<cc-env-var-create>`: add a strict mode